requirements.txt
asgiref==3.4.1
Django==4.0
psycopg2-binary>=2.8
django-cors-headers>=3.10
djangorestframework>=3.13
python-gitlab